# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Imari API
  version: 0.1.0
  description: Personalised financial literacy platform — ALU pilot.
servers:
  - url: https://api.imari.app/v1
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  schemas:
...
    Invite:
      type: object
      properties:
        code: { type: string, example: ALU-PILOT }
        email: { type: [string, 'null'] }
        maxUses: { type: [integer, 'null'] }
        uses: { type: integer }
        expiresAt: { type: string, format: date-time }
        status: { type: string, enum: [active, used, expired, revoked] }
...
paths:
...
  /modules/{slug}:
    get:
      parameters: [{ name: slug, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: Module with lessons, activity and quiz (answers stripped) } }
...
  /admin/invites:
    post:
      requestBody: { content: { application/json: { schema: { type: object, properties: { email: { type: string }, note: { type: string }, maxUses: { type: [integer, 'null'] }, days: { type: integer } } } } } }
      responses: { '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Invite' } } } } }
  /admin/invites/{id}:
    patch:
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { action: { type: string, enum: [revoke, extend] }, days: { type: integer } } } } } }
      responses: { '200': { description: Updated } }
  /admin/credentials/{id}:
    patch:
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      requestBody: { content: { application/json: { schema: { type: object, properties: { status: { type: string, enum: [valid, revoked] }, reason: { type: string } } } } } }
      responses: { '200': { description: Updated on DB and chain } }
...